Fix and break out unit tests.
