+++ projects/packages/scheduled-updates/src/wpcom-endpoints/class-wpcom-rest-api-v2-endpoint-update-schedules-capabilities.php
@@ -33,7 +33,8 @@ class WPCOM_REST_API_V2_Endpoint_Update_Schedules_Capabilities extends WP_REST_C
 	 * WPCOM_REST_API_V2_Endpoint_Atomic_Hosting_Update_Schedule constructor.
 	 */
 	public function __construct() {
-		add_action( 'rest_api_init', array( $this, 'register_routes' ) );
+		// Priority 9 to run before the main endpoint and avoid "/capabilities" being treated as a schedule ID.
+		add_action( 'rest_api_init', array( $this, 'register_routes' ), 9 );
 	}
 
 	/**
