+++ projects/packages/scheduled-updates/tests/lib/functions-wordpress.php
@@ -5,6 +5,25 @@
  * @package automattic/scheduled-updates
  */
 
+if ( ! function_exists( 'wpcom_rest_api_v2_load_plugin_files' ) ) {
+	/**
+	 * Loop through endpoint files and load them.
+	 *
+	 * @param string $file_pattern Path pattern to the endpoints (pattern must be supported by glob()).
+	 */
+	function wpcom_rest_api_v2_load_plugin_files( $file_pattern ) {
+		$plugins = glob( __DIR__ . '/' . $file_pattern );
+
+		if ( ! is_array( $plugins ) ) {
+			return;
+		}
+
+		foreach ( array_filter( $plugins, 'is_file' ) as $plugin ) {
+			require_once $plugin;
+		}
+	}
+}
+
 if ( ! function_exists( 'wpcom_rest_api_v2_load_plugin' ) ) {
 	/**
 	 * A drop-in for a WordPress.com function.
