+++ projects/packages/scheduled-updates/tests/php/class-wpcom-rest-api-v2-endpoint-update-schedules-test.php
@@ -754,205 +754,4 @@ class WPCOM_REST_API_V2_Endpoint_Update_Schedules_Test extends \WorDBless\BaseTe
 
 		$this->assertSame( $expected_result, get_option( 'auto_update_plugins' ) );
 	}
-
-	/**
-	 * Make sure unauthorized users can't get in to capabilities.
-	 *
-	 * @covers ::get_capabilities
-	 */
-	public function test_non_admin_user_capabilities() {
-		$request = new WP_REST_Request( 'GET', '/wpcom/v2/update-schedules/capabilities' );
-		$result  = rest_do_request( $request );
-
-		$this->assertSame( 401, $result->get_status() );
-	}
-
-	/**
-	 * Make sure authorized users can see data for capabilities
-	 *
-	 * @covers ::get_capabilities
-	 */
-	public function test_admin_user_capabilities() {
-		$request = new WP_REST_Request( 'GET', '/wpcom/v2/update-schedules/capabilities' );
-		wp_set_current_user( $this->admin_id );
-		$result = rest_do_request( $request );
-
-		$this->assertSame( 200, $result->get_status() );
-	}
-
-	/**
-	 * Test adding a log entry for a non-existent schedule.
-	 *
-	 * @covers ::add_log
-	 */
-	public function test_add_log_invalid_schedule() {
-		wp_set_current_user( $this->admin_id );
-
-		$request = new WP_REST_Request( 'PUT', '/wpcom/v2/update-schedules/' . Scheduled_Updates::generate_schedule_id( array() ) . '/logs' );
-		$request->set_body_params(
-			array(
-				'action'  => Scheduled_Updates_Logs::PLUGIN_UPDATES_START,
-				'message' => 'Starting plugin updates.',
-			)
-		);
-		$result = rest_do_request( $request );
-
-		$this->assertSame( 404, $result->get_status() );
-		$this->assertEmpty( get_option( Scheduled_Updates::PLUGIN_CRON_HOOK ) );
-	}
-
-	/**
-	 * Test retrieving logs
-	 *
-	 * @covers ::add_log
-	 */
-	public function test_get_logs() {
-		wp_set_current_user( $this->admin_id );
-
-		$schedule_id = $this->create_test_schedule();
-
-		$request = new WP_REST_Request( 'GET', '/wpcom/v2/update-schedules/' . $schedule_id . '/logs' );
-		$result  = rest_do_request( $request );
-
-		$this->assertSame( 200, $result->get_status() );
-		$this->assertSame( array(), $result->get_data() );
-
-		$sync_option = get_option( Scheduled_Updates::PLUGIN_CRON_HOOK );
-		$this->assertIsArray( $sync_option );
-		$this->assertNull( $sync_option[ $schedule_id ]->last_run_timestamp );
-		$this->assertNull( $sync_option[ $schedule_id ]->last_run_status );
-	}
-
-	/**
-	 * Test adding a log entry and retrieving it
-	 *
-	 * @covers ::add_log
-	 * @covers ::get_logs
-	 */
-	public function test_add_and_get_log() {
-		wp_set_current_user( $this->admin_id );
-
-		$schedule_id = $this->create_test_schedule();
-
-		$request = new WP_REST_Request( 'PUT', '/wpcom/v2/update-schedules/' . $schedule_id . '/logs' );
-		$request->set_body_params(
-			array(
-				'action'  => Scheduled_Updates_Logs::PLUGIN_UPDATES_START,
-				'message' => 'Starting plugin updates.',
-			)
-		);
-		$result = rest_do_request( $request );
-
-		$this->assertSame( 200, $result->get_status() );
-
-		$request = new WP_REST_Request( 'GET', '/wpcom/v2/update-schedules/' . $schedule_id . '/logs' );
-		$result  = rest_do_request( $request );
-
-		$this->assertSame( 200, $result->get_status() );
-		$this->assertCount( 1, $result->get_data() );
-		$this->assertSame( Scheduled_Updates_Logs::PLUGIN_UPDATES_START, $result->get_data()[0][0]['action'] );
-
-		$sync_option = get_option( Scheduled_Updates::PLUGIN_CRON_HOOK );
-		$this->assertIsArray( $sync_option );
-		$this->assertNull( $sync_option[ $schedule_id ]->last_run_timestamp );
-		$this->assertSame( 'in-progress', $sync_option[ $schedule_id ]->last_run_status );
-	}
-
-	/**
-	 * Test adding multiple runs and retrieving them
-	 *
-	 * @covers ::add_log
-	 * @covers ::get_logs
-	 */
-	public function test_add_and_get_multiple_logs() {
-		wp_set_current_user( $this->admin_id );
-
-		$schedule_id = $this->create_test_schedule();
-
-		// Simulate 5 runs.
-		for ( $i = 0;$i < 5;$i++ ) {
-			$request = new WP_REST_Request( 'PUT', '/wpcom/v2/update-schedules/' . $schedule_id . '/logs' );
-			$request->set_body_params(
-				array(
-					'action'  => Scheduled_Updates_Logs::PLUGIN_UPDATES_START,
-					'message' => 'Starting plugin updates.',
-				)
-			);
-			$result = rest_do_request( $request );
-			$this->assertSame( 200, $result->get_status() );
-
-			$request = new WP_REST_Request( 'PUT', '/wpcom/v2/update-schedules/' . $schedule_id . '/logs' );
-			$request->set_body_params(
-				array(
-					'action'  => Scheduled_Updates_Logs::PLUGIN_UPDATES_SUCCESS,
-					'message' => 'Ending plugin updates.',
-				)
-			);
-			$result = rest_do_request( $request );
-			$this->assertSame( 200, $result->get_status() );
-		}
-
-		$request = new WP_REST_Request( 'GET', '/wpcom/v2/update-schedules/' . $schedule_id . '/logs' );
-		$result  = rest_do_request( $request );
-
-		$this->assertSame( 200, $result->get_status() );
-		$this->assertCount( Scheduled_Updates_Logs::MAX_RUNS_PER_SCHEDULE, $result->get_data() );
-		$this->assertSame( Scheduled_Updates_Logs::PLUGIN_UPDATES_START, $result->get_data()[0][0]['action'] );
-
-		$sync_option = get_option( Scheduled_Updates::PLUGIN_CRON_HOOK );
-		$this->assertNotNull( $sync_option[ $schedule_id ]->last_run_timestamp );
-		$this->assertSame( 'success', $sync_option[ $schedule_id ]->last_run_status );
-	}
-
-	/**
-	 * Test adding a log when unauthorized
-	 *
-	 * @covers ::add_log
-	 */
-	public function test_add_log_unauthorized() {
-		$request = new WP_REST_Request( 'PUT', '/wpcom/v2/update-schedules/' . Scheduled_Updates::generate_schedule_id( array() ) . '/logs' );
-		$request->set_body_params(
-			array(
-				'action'  => Scheduled_Updates_Logs::PLUGIN_UPDATES_START,
-				'message' => 'Starting plugin updates.',
-			)
-		);
-		$result = rest_do_request( $request );
-		$this->assertSame( 401, $result->get_status() );
-	}
-
-	/**
-	 * Create schedule
-	 *
-	 * @param int $i Schedule index.
-	 */
-	private function create_test_schedule( $i = 0 ) {
-		$request           = new \WP_REST_Request( 'POST', '/wpcom/v2/update-schedules' );
-		$scheduled_plugins = array( 'test/test' . $i . '.php' );
-		$request->set_body_params(
-			array(
-				'plugins'  => $scheduled_plugins,
-				'schedule' => $this->get_schedule( "next Monday {$i}:00" ),
-			)
-		);
-
-		$result = rest_do_request( $request );
-		return $result->get_data();
-	}
-
-	/**
-	 * Get a schedule.
-	 *
-	 * @param string $timestamp Schedule timestamp.
-	 * @param string $interval Schedule interval.
-	 * @param array  $health_check_paths Health check paths.
-	 * @return array
-	 */
-	private function get_schedule( $timestamp = 'next Monday 8:00', $interval = 'weekly', $health_check_paths = array() ) {
-		return array(
-			'timestamp'          => strtotime( $timestamp ),
-			'interval'           => $interval,
-			'health_check_paths' => $health_check_paths,
-		);
-	}
 }
